Trait rust_sparseglm::estimators::traits::Fit

pub trait Fit<DM: DesignMatrix, T: AsMultiTargets, E: Error> {
    type Object;
    fn fit(&self, dataset: &DatasetBase<DM, T>) -> Result<Self::Object, E>;
}

Fit trait

The fittable trait allows an estimator to be fitted to a dataset (a combination of design matrix and targets). More formally, the model estimates coefficients that minimizes an empirical risk (loss function).

Associated Types

type Object

Required methods

fn fit(&self, dataset: &DatasetBase<DM, T>) -> Result<Self::Object, E>