Trait rust_sparseglm::Float

pub trait Float: FromPrimitive + Float + PartialOrd + Sync + Send + Default + Display + Debug + Signed + Sum + NumAssignOps + AsPrimitive<usize> + for<'a> AddAssign<&'a Self> + for<'a> MulAssign<&'a Self> + for<'a> SubAssign<&'a Self> + for<'a> DivAssign<&'a Self> + MulAdd<Output = Self> + ScalarOperand + AbsDiffEq {
    type Lapack: Float;
    fn cast<T: NumCast>(x: T) -> Self { ... }
}

Float point numbers

This trait bound multiplexes to the most common assumption of floating point number and implement them for 32bit and 64bit float points. Ref: <https://github.com/rust-ml/linfa/blob/master/src/dataset/mod.rs#L36>

Associated Types

type Lapack: Float

Provided methods

fn cast<T: NumCast>(x: T) -> Self